deduplicate utils
