+++ src/napari_threedee/annotators/paths/annotator.py
@@ -11,7 +11,7 @@ from napari_threedee.annotators.paths.constants import (
     PATH_COLOR_FEATURES_KEY,
     COLOR_CYCLE,
 )
-from napari_threedee.mouse_callbacks import add_point_on_plane
+from napari_threedee.utils.mouse_callbacks import add_point_on_plane
 from napari_threedee.utils.napari_utils import add_mouse_callback_safe, \
     remove_mouse_callback_safe
 from napari_threedee.annotators.constants import N3D_METADATA_KEY, ANNOTATION_TYPE_KEY
