+++ src/napari_threedee/annotators/points/annotator.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from napari_threedee._backend.threedee_model import N3dComponent
 from napari_threedee.annotators.points.validation import validate_layer
-from napari_threedee.mouse_callbacks import add_point_on_plane
+from napari_threedee.utils.mouse_callbacks import add_point_on_plane
 from napari_threedee.utils.napari_utils import add_mouse_callback_safe, \
     remove_mouse_callback_safe
